+++ src/core/leoCommands.ts
@@ -50,6 +50,10 @@ import { HelpCommandsClass } from '../commands/helpCommands';
 import { KillBufferCommandsClass } from '../commands/killBufferCommands';
 import { RectangleCommandsClass } from '../commands/rectangleCommands';
 import * as typescript from 'typescript';
+import * as difflib from 'difflib';
+import * as csv from 'csvtojson';
+import KSUID = require('ksuid');
+
 const dayjs = require('dayjs');
 const utc = require('dayjs/plugin/utc');
 dayjs.extend(utc);
@@ -1252,13 +1256,28 @@ export class Commands {
             ? { c: c, g: g, input: '', p: p }
             : {};
 
-        d['Buffer'] = Buffer;
+        // * VSCode API
         d['vscode'] = vscode;
+
+        // * Default NODE Modules
+        d['Buffer'] = Buffer;
         d['crypto'] = crypto;
         d['os'] = os;
         d['path'] = path;
         d['process'] = process;
         d['child_process'] = child_process;
+
+        // * Imported Libraries
+        d['SQL'] = g.SQL;
+        d['pako'] = g.pako;
+        d['showdown'] = g.showdown;
+        d['JSZip'] = g.JSZip;
+        d['dayjs'] = g.dayjs;
+        d['md5'] = g.md5;
+        d['csvtojson'] = csv;
+        d['difflib'] = difflib;
+        d['elementtree'] = et;
+        d['ksuid'] = KSUID;
 
         if (define_name) {
             d['__name__'] = define_name;